requirements.txt
torch>=2.1
transformers>=4.39
biopython>=1.83
numpy>=1.24
tqdm>=4.66
